chart1: Use udev enumerated serial ports (#1964).
If libudev is available, use it to enumerate the available ports.
Since libudev is ubiquoitous, this works fine for basically all native
linux hosts.  It also solves the problem with udev device links, since
also these are enumerated.

However, it does not solve the problem for flatpak, which
does not support libudev in the runtime.