Add sysfs-based serial port enumeration.
Adds a new OCPN_USE_SYSFS_PORTS configuration option. If enabled it
uses the udev interface under /sys, notably /sys/class/tty, to enumerate
all available serial devices.

Device symlinks are included. ttyS0 and ttyS1 are not included unless
they can be handled using ioctl() -- this might need to be revised.

The primary usecase for this is flatpak, where libudev is not available.
This requires that the sandbox has read-only access to /sys which should be
possible even without the devel permissions which are applied today.