model: Ensure that outgoing messages have empty source
The source part of NavMsg is intended for incoming traffic. Setting
this to an empty address on outgoing traffic is logical and makes
it possible to determine if a message on the bus is incoming our
outgoing.

Furthermore, for the n0183 driver the dest address os moot and could
as well be empty.