signalk net driver: Dont add crlf - #4889
The driver was adding a trailing \r\n to the data sent to uppper layers.
This is plain wrong, these delimiters are a transport thing. If
anything depends on it it should be fixed, the correct thing for a
driver is to forward data without transport delimiters.

Remove the \r\n from data forwarded to upper layers.

While on it, fix a bug condition caused by using a raw pointer while
the shared_ptr holding the object is out of scope, an obvious crash
risk.

Closes: #4889